+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 1.26.0-beta.x
 
-- Update docs for the new `system.accuntNextIndex` (Thanks to https://github.com/shawntabrizi)
+- Update docs for the new `system.accountNextIndex` (Thanks to https://github.com/shawntabrizi)
 - Adjust votingBalance to only return `freeBalance` (only council voting is free + reserved)
 - Additional checks for referenced Vec sub-type deconstruction in typegen
+- Allow for API `typesBundle` option to inject a group of chain versions types
 - `@polkadot/util` 3.0
 
 ## 1.25.1 Jul 20, 2020